requirements.txt
elasticsearch>=7.10.1
PyYAML>=5.3.1
mistletoe>=0.7.2
tqdm>=4.55.1
pandas>=1.1.5
jupyter>=1.0.0
matplotlib>=3.3.3
seaborn>=0.11.1
coverage>=5.3.1
coveralls>=3.0.0

# for doc-tests
Sphinx>=3.4.3
sphinx-rtd-theme>=0.5.1
docutils>=0.16
